requirements.txt
python-dotenv
langchain
openai
chromadb
pypdf
tiktoken
faiss-cpu
streamlit
pyyaml